tests/requirements.txt
# Tests requirements
mock
mitmproxy==0.10.1
netlib==0.10.1
pytest-twisted
pytest-cov
jmespath
testfixtures
# optional for shell wrapper tests
bpython
ipython
